Code review: rename custom hooks file from markers.js to ramp-hooks.js
