+++ src/components/MarkersDisplay/MarkerUtils/CreateMarker.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { parseMarkerAnnotation } from '@Services/playlist-parser';
 import { validateTimeInput, timeToS, timeToHHmmss } from '@Services/utility-helpers';
 import { SaveIcon, CancelIcon } from '@Services/svg-icons';
-import { usePlayer } from '@Services/markers';
+import { usePlayer } from '@Services/ramp-hooks';
 
 const CreateMarker = ({ newMarkerEndpoint, canvasId, handleCreate, csrfToken }) => {
   const [isOpen, setIsOpen] = React.useState(false);
